# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Diwa",
        "description": "Diwa is an unofficial simple API from [distrowatch.com](https://distrowatch.com/)",
        "contact": {
            "email": "mzulfahmi807@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/Zzzul/diwa/blob/main/LICENSE"
        },
        "version": "2.0"
    },
    "paths": {
...
        "/api/v2/news/{id}": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Get Distribution News information detail",
                "description": "If {news_id} not found, distrowatch.com will return the home page. make sure {news_id} is correct",
                "operationId": "getDistributionNewsById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "News Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "11302"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
        "/api/v2/filter/news": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Get specific distribution news",
                "description": "If one of the {params} not found, distrowatch.com will return the home page with default params(all). make sure all {params} are correct",
                "operationId": "FilterDistributionNews",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Distribution Name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "ubuntu"
                    },
                    {
                        "name": "release",
                        "in": "query",
                        "description": "Release Version",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "stable"
                    },
                    {
                        "name": "month",
                        "in": "query",
                        "description": "Month",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "all"
                    },
                    {
                        "name": "year",
                        "in": "query",
                        "description": "Year",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "2022"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        }
    },
...
}